refactor: move year utils
