+++ lib/abcApi.ts
@@ -1,27 +1,12 @@
 "use server"
 
-// October is designated as the cut-off month for determining the current academic year
-const ROLLOVER_MONTH = 9
+import { currentShortYear } from "@/lib/util/academicYear"
 
 const ABC_ROOT = process.env.ABC_API || "https://abc-api.doc.ic.ac.uk/"
 const ABC_USERNAME = process.env.API_ROLE_USERNAME || "adumble"
 const ABC_PASSWORD = process.env.API_ROLE_PASSWORD || "bob"
 
 const abcCurrentYear = `${ABC_ROOT}/${currentShortYear()}`
-
-/**
- * Returns the current academic year as a short string using today's date: e.g. 14 Feb 2025 -> '2425'
- * The rollover month (i.e. using previous or future year) is a constant.
- *
- * @returns {string} The short year string.
- */
-function currentShortYear(): string {
-  const date = new Date()
-  const year = date.getFullYear()
-  const month = date.getMonth()
-  const [curr, prev, next] = [year, year - 1, year + 1].map(y => `${y}`.slice(2))
-  return month < ROLLOVER_MONTH ? `${prev}${curr}` : `${curr}${next}`
-}
 
 async function fetchFromAbc(endpoint: string, xProxiedUser?: string) {
   const credentials = Buffer.from(`${ABC_USERNAME}:${ABC_PASSWORD}`).toString("base64")
